requirements.txt
numpy
Pillow
python-multipart
torch
torchvision